data-processing/requirement.txt
PyMySQL
PyYAML
autopep8
beautifulsoup4
requests
memory_profiler
pandas
